FFM-9461 remove old code
**What**

- Removes a lot of the old proxy code that we won't be using

**Why**

- To try and give us some sort of cleanish slate to start the new work
